refactor: extract Workflow Defaults + Global Hotkeys settings panes (#527, #355)
Continues the SettingsView pane extraction (A1 Diagnostics/Privacy #524,
A2 Audio/Permissions #526). Moves the "Workflow Defaults" and "Global
Hotkeys" GroupBoxes into Views/Settings/SettingsGeneralPanes.swift as
SettingsWorkflowDefaultsPane and SettingsGlobalHotkeysPane, rendered in
their existing body positions so the layout is unchanged.

Their exclusive helpers move with them: calloutColor (Workflow Defaults)
and hotkeyRow (Global Hotkeys). With their only callers gone, both are
removed from SettingsView.

Pure UI relocation: no field added/removed/rewired, every setting keeps its
SettingsStore key, no controller/validation change. Build clean (no new
warnings); full unit suite green; all 7 BugNarratorSettingsUITests pass.

Closes #527
Refs #355
